Move plot to own module
