+++ src/shiver/views/workspace_tables.py
@@ -18,15 +18,13 @@ from qtpy.QtWidgets import (
 
 from qtpy.QtCore import Qt, QSize, Signal
 from qtpy.QtGui import QIcon, QPixmap, QCursor
-
-import matplotlib.pyplot as plt
-from mantidqt.widgets.sliceviewer.presenters.presenter import SliceViewer
-from mantidqt.plotting.functions import manage_workspace_names, plot_md_ws_from_names
+from mantidqt.plotting.functions import plot_md_ws_from_names
 
 from shiver.views.sample import SampleView
 from shiver.presenters.sample import SamplePresenter
 from shiver.models.sample import SampleModel
 from .histogram_parameters import INVALID_QLISTWIDGET
+from .plots import do_colorfill_plot, do_slice_viewer
 
 Frame = Enum("Frame", {"None": 1000, "QSample": 1001, "QLab": 1002, "HKL": 1003})
 
@@ -654,20 +652,3 @@ def get_icon(name: str) -> QIcon:
         )
 
     raise ValueError(f"{name} doesn't correspond to a valid icon")
-
-
-@manage_workspace_names
-def do_colorfill_plot(workspaces):
-    """Create a colormesh plot for the provided workspace"""
-    fig, axis = plt.subplots(subplot_kw={"projection": "mantid"})
-    colormesh = axis.pcolormesh(workspaces[0])
-    axis.set_title(workspaces[0].name())
-    fig.colorbar(colormesh)
-    fig.show()
-
-
-@manage_workspace_names
-def do_slice_viewer(workspaces, parent=None):
-    """Open sliceviewer for the provided workspace"""
-    presenter = SliceViewer(ws=workspaces[0], parent=parent)
-    presenter.view.show()
